Add wrapper around languagetool
Add a simple wrapper around languagetool that pipes the current buffer
to languagetool and collect the resulting json. The json data is then
inspected, and all errors are syntax highlighted and a virtual-text
property added to show the issues found by the tool.

As languagetool is slow to run there is no automatic update of the
grammar linting and the plugin depends on the user to either issue a new
check or rest the view once the issues have been examined.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>