Add configuration to ignore lines
For some file types it can be useful to ignore certain lines, for
example when replying to an email ignoring lines that starts with "> "
can be useful.

To allow for this a new configuration variable can be set to ignore
lines that match the pattern. For example with the mail file type,

    let g:rtgram_markers_ignore= ["^>"]

Or the gitcommit file type to ignore lines that are comments,

    let g:rtgram_markers_ignore= ["^#"]

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>