Use vim job to run languagetool in background
Instead of blocking the vim process while languagetool runs use vim job
control to run it in the background. This avoids vim freezing up while
the check is running.

This is also a better design to build upon if the plugin shall ever
evolve to automatically run checks as the text is modified. Or if the
plugin learns how to make use of languagetool running in a server mode
to avoid the cost of starting spawning a new process for each check.

The drawback is that the user can start a new check before the first one
is completed resulting in multiple virtual-text entries for the same
issues being reported. This drawback could be fixed by some job control
logic only allowing a single job at a time but since running a check in
the background already races with the user modifying the buffer before
the job completes there is little point in doing so.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>